api action完善
